+++ maintainerr/CHANGELOG.md
@@ -1,6 +1,6 @@
 
 ## 3.2.0-2 (2026-03-31)
-- Fix configuration lost after container restart by using persistent addon_config directory
+- Fix configuration lost after container restart by symlinking /opt/data to persistent /config directory
 
 ## 3.2.0 (2026-03-28)
 - Update to latest version from maintainerr/maintainerr (changelog : https://github.com/maintainerr/maintainerr/releases)